Restructure include directories into hmat_lib and internal
This way the header files that will be installed with the library are isolated from the header files that are only required inside the library
