+++ include/hmat_lib/error.h
