+++ src/constructors.c
@@ -1,9 +1,9 @@
 #include <stdlib.h>
 
-#include "../include/lapack_wrapper.h"
-#include "../include/blas_wrapper.h"
-#include "../include/hodlr.h"
-#include "../include/error.h"
+#include "../include/internal/lapack_wrapper.h"
+#include "../include/internal/blas_wrapper.h"
+#include "../include/hmat_lib/hodlr.h"
+#include "../include/hmat_lib/error.h"
 
 
 /**
