+++ tests/Makefile
@@ -1,7 +1,8 @@
 SRC_DIR := ../src
 TESTS_DIR := src
 UTIL_INCLUDE_DIR := include
-INCLUDE_DIR := ../include
+INCLUDE_DIR := ../include/hmat_lib
+INTERNAL_INCLUDE_DIR := ../include/internal
 DEV_DIR := ../dev
 BUILD_DIR := build
 OUT_DIR := ../bin
@@ -36,7 +37,7 @@ test_hodlr_algebra.o: ${TESTS_DIR}/test_hodlr_hodlr_algebra.c
 		-c ${TESTS_DIR}/test_hodlr_hodlr_algebra.c \
 		-o ${BUILD_DIR}/test_hodlr_algebra.o \
 		-lopenblas -lm -lcriterion \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${WARNINGS} \
 		${EXTRA_FLAGS}
 
@@ -58,7 +59,7 @@ test_dense_algebra.o: ${TESTS_DIR}/test_hodlr_dense_algebra.c lapack_wrapper.o u
 		-c ${TESTS_DIR}/test_hodlr_dense_algebra.c \
 		-o ${BUILD_DIR}/test_dense_algebra.o \
 		-lopenblas -lm -lcriterion \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${WARNINGS} \
 		${EXTRA_FLAGS}
 
@@ -79,7 +80,7 @@ test_vector_algebra.o: ${TESTS_DIR}/test_vector_algebra.c lapack_wrapper.o utils
 		-c ${TESTS_DIR}/test_vector_algebra.c \
 		-o ${BUILD_DIR}/test_vector_algebra.o \
 		-lopenblas -lm -lcriterion -lmimick \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${EXTRA_FLAGS}
 
 
@@ -100,7 +101,7 @@ test_tree_construct.o: ${TESTS_DIR}/test_tree_construct.c
 		-c ${TESTS_DIR}/test_tree_construct.c \
 		-o ${BUILD_DIR}/test_tree_construct.o \
 		-lopenblas \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${WARNINGS} \
 		${EXTRA_FLAGS}
 
@@ -117,7 +118,7 @@ test_allocators.o: ${TESTS_DIR}/test_allocators.c
 	gcc \
 		-c ${TESTS_DIR}/test_allocators.c \
 		-o ${BUILD_DIR}/test_allocators.o \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${EXTRA_FLAGS}
 
 test_real_data: test_real_data.o utils.o common_data.o io.o allocators.o constructors_omp.o lapack_wrapper.o vector_algebra.o dense_algebra.o common.o hodlr_algebra.o main_utils.o
@@ -136,7 +137,7 @@ test_real_data: test_real_data.o utils.o common_data.o io.o allocators.o constru
 		${BUILD_DIR}/hodlr_algebra.o \
 		-o ${OUT_DIR}/test_real_data \
 		-lopenblas -lm -lcriterion \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${WARNINGS} ${EXTRA_FLAGS} \
 		-fopenmp \
 		${REAL_DATA_PRINT_S}
@@ -145,7 +146,7 @@ test_real_data.o: ${TESTS_DIR}/test_real_data.c
 	gcc \
 		-c ${TESTS_DIR}/test_real_data.c \
 		-o ${BUILD_DIR}/test_real_data.o \
-		-I ${INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} -I ${UTIL_INCLUDE_DIR} \
 		${EXTRA_FLAGS} ${REAL_DATA_PRINT_S} -fopenmp
 
 
@@ -157,19 +158,22 @@ allocators.o: ${SRC_DIR}/allocators.c
 	${COMPILER} \
 		-c ${SRC_DIR}/allocators.c \
 		-o ${BUILD_DIR}/allocators.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS}
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS}
 
 constructors.o: ${SRC_DIR}/constructors.c
 	${COMPILER} \
 		-c ${SRC_DIR}/constructors.c \
 		-o ${BUILD_DIR}/constructors.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS} 
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS} 
 
 constructors_omp.o: ${SRC_DIR}/constructors.c
 	${COMPILER} \
 		-c ${SRC_DIR}/constructors.c \
 		-o ${BUILD_DIR}/constructors_omp.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS} \
 		-fopenmp
 
 lapack_wrapper.o: ${SRC_DIR}/lapack_wrapper.c
@@ -177,26 +181,29 @@ lapack_wrapper.o: ${SRC_DIR}/lapack_wrapper.c
 		-c ${SRC_DIR}/lapack_wrapper.c \
 		-o ${BUILD_DIR}/lapack_wrapper.o \
 		-lopenblas \
-		-I ${INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
 		${EXTRA_FLAGS}
 
 vector_algebra.o: ${SRC_DIR}/vector_algebra.c
 	${COMPILER} \
 		-c ${SRC_DIR}/vector_algebra.c \
 		-o ${BUILD_DIR}/vector_algebra.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS}
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS}
 
 dense_algebra.o: ${SRC_DIR}/dense_algebra.c
 	${COMPILER} \
 		-c ${SRC_DIR}/dense_algebra.c \
 		-o ${BUILD_DIR}/dense_algebra.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS}
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS}
 
 hodlr_algebra.o: ${SRC_DIR}/hodlr_algebra.c
 	${COMPILER} \
 		-c ${SRC_DIR}/hodlr_algebra.c \
 		-o ${BUILD_DIR}/hodlr_algebra.o \
-		-I ${INCLUDE_DIR} ${WARNINGS} ${EXTRA_FLAGS}
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
+		${WARNINGS} ${EXTRA_FLAGS}
 
 
 
@@ -235,7 +242,7 @@ main_utils.o: ${SRC_DIR}/utils.c
 	gcc \
 		-c ${SRC_DIR}/utils.c \
 		-o ${BUILD_DIR}/main_utils.o \
-		-I ${INCLUDE_DIR} \
+		-I ${INCLUDE_DIR} -I ${INTERNAL_INCLUDE_DIR} \
 		${EXTRA_FLAGS}
